requirements.txt
alembic==1.8.1
anyio==3.6.1
asyncpg==0.26.0
cfgv==3.3.1
click==8.1.3
colorama==0.4.5
databases==0.6.0
distlib==0.3.5
fastapi==0.78.0
fastapi-jwt-auth==0.5.0
filelock==3.7.1
greenlet==1.1.2
gunicorn==20.1.0
h11==0.13.0
identify==2.5.2
idna==3.3
Mako==1.2.1
MarkupSafe==2.1.1
nodeenv==1.7.0
passlib==1.7.4
platformdirs==2.5.2
pre-commit==2.20.0
psycopg2-binary==2.9.3
pydantic==1.9.1
PyJWT==1.7.1
python-decouple==3.6
PyYAML==6.0
sniffio==1.2.0
SQLAlchemy==1.4.39
starlette==0.19.1
toml==0.10.2
typing_extensions==4.3.0
uvicorn==0.18.2
virtualenv==20.16.0
